Update cloudfunctions
* Refactor it into multiple files
* Create search function
